feat: transform renderMardown into a component
