+++ app/posts/[slug]/page.tsx
@@ -1,10 +1,10 @@
 import { notFound } from 'next/navigation'
 
 import { FeaturedImage } from '@/components/atoms/FeaturedImage'
+import { MarkdownRenderer } from '@/components/organisms/MarkdownRenderer'
 import { PostArchives } from '@/components/organisms/PostArchives'
 import { PostHeader } from '@/components/organisms/PostHeader'
 import { client } from '@/lib/clients'
-import { renderAndSanitizeMarkdown } from '@/lib/renderMarkdown'
 import { getImageUrl } from '@/lib/utils'
 
 export const generateStaticParams = async () => {
@@ -44,8 +44,6 @@ const Page = async ({ params }: { params: Promise<{ slug: string }> }) => {
 
     const { featuredImage, content, title, tags } = post
 
-    const renderedContent = renderAndSanitizeMarkdown(content)
-
     const MAX_RELATED_POSTS = 3
 
     const relatedPosts = tags?.length
@@ -64,11 +62,7 @@ const Page = async ({ params }: { params: Promise<{ slug: string }> }) => {
             )}
 
             <article className="container prose mx-auto max-w-6xl px-6 dark:prose-invert lg:prose-xl">
-                {renderedContent && (
-                    <div
-                        dangerouslySetInnerHTML={{ __html: renderedContent }}
-                    />
-                )}
+                <MarkdownRenderer content={content} />
             </article>
 
             {relatedPosts.length !== 0 && (
